moving more moving more eliza code around, placing models in their own package. not fully done yet
