+++ kifi-backend/modules/eliza/app/com/keepit/common/cache/ElizaCacheModule.scala
@@ -5,7 +5,8 @@ import com.keepit.common.logging.AccessLog
 import com.google.inject.{Provides, Singleton}
 import com.keepit.model._
 import com.keepit.social.BasicUserUserIdCache
-import com.keepit.eliza.{MessagesForThreadIdCache, MessageThreadExternalIdCache}
+import com.keepit.eliza._
+import com.keepit.eliza.model._
 import com.keepit.search.ActiveExperimentsCache
 import com.keepit.common.usersegment.UserSegmentCache
