+++ kifi-backend/modules/eliza/app/com/keepit/eliza/ElizaThreadInfo.scala
@@ -6,6 +6,7 @@ import com.keepit.common.db._
 import com.keepit.common.time._
 import com.keepit.model._
 import com.keepit.social.{BasicUserLikeEntity, BasicUser}
+import com.keepit.eliza.model._
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
