+++ kifi-backend/modules/eliza/app/com/keepit/eliza/MessageWithBasicUser.scala
@@ -8,6 +8,7 @@ import com.keepit.common.db.ExternalId
 import com.keepit.common.time._
 import com.keepit.model._
 import com.keepit.social.{BasicUserLikeEntity, BasicUser}
+import com.keepit.eliza.model._
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
