+++ kifi-backend/modules/eliza/app/com/keepit/eliza/MessagingAnalytics.scala
@@ -11,6 +11,7 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import com.keepit.shoebox.ShoeboxServiceClient
 import com.keepit.realtime.PushNotification
 import com.keepit.common.db.slick.Database
+import com.keepit.eliza.model._
 
 @Singleton
 class MessagingAnalytics @Inject() (
