+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/MessagingTypeMappers.scala
@@ -1,4 +1,4 @@
-package com.keepit.eliza
+package com.keepit.eliza.model
 
 import scala.slick.lifted.{BaseTypeMapper}
 import com.keepit.common.db.{Id, ExternalId}
@@ -28,16 +28,16 @@ object MessagingTypeMappers {
 
   implicit object MessageThreadParticipantsMapper extends BaseTypeMapper[MessageThreadParticipants] {
     def apply(profile: BasicProfile) = new StringMapperDelegate[MessageThreadParticipants](profile) {
-      
+
       def safeDestToSource(source: String): MessageThreadParticipants = {
         Json.parse(source).validate[MessageThreadParticipants] match {
           case JsSuccess(mtps,_) => mtps
           case _ => throw InvalidDatabaseEncodingException(s"Could not decode JSON for MessageThreadParticipants: $source")
         }
       }
-      
+
       def sourceToDest(dest: MessageThreadParticipants): String = Json.stringify(Json.toJson(dest))
-      
+
       def zero: MessageThreadParticipants = MessageThreadParticipants(Set[Id[User]]())
 
     }
@@ -45,7 +45,7 @@ object MessagingTypeMappers {
 
   implicit object NormalizedUriIdSeqMapper extends BaseTypeMapper[Seq[Id[NormalizedURI]]] {
     def apply(profile: BasicProfile) = new StringMapperDelegate[Seq[Id[NormalizedURI]]](profile) {
-      
+
       implicit val nUriIdFormat = Id.format[NormalizedURI]
 
       def safeDestToSource(source: String): Seq[Id[NormalizedURI]] = {
@@ -56,11 +56,11 @@ object MessagingTypeMappers {
           case _ => throw InvalidDatabaseEncodingException(s"Could not decode JSON for Seq of Normalized URI ids: $source")
         }
       }
-      
+
       def sourceToDest(dest: Seq[Id[NormalizedURI]]): String = {
         Json.stringify(JsArray(dest.map( x => JsNumber(x.id) )))
       }
-      
+
       def zero: Seq[Id[NormalizedURI]] = Seq[Id[NormalizedURI]]()
     }
   }
