+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/NonUserThread.scala
@@ -1,0 +1,80 @@
+package com.keepit.eliza.model
+
+import com.google.inject.{Inject, Singleton, ImplementedBy}
+import com.keepit.common.db.slick.{StringMapperDelegate, Repo, DbRepo, DataBaseComponent}
+import com.keepit.common.db.slick.FortyTwoTypeMappers._
+import com.keepit.common.db.slick.DBSession.{RWSession, RSession}
+import org.joda.time.DateTime
+import com.keepit.common.logging.Logging
+import com.keepit.common.time._
+import com.keepit.common.db.{States, State, Model, Id}
+import com.keepit.model.{EContact, User, NormalizedURI}
+import play.api.libs.json._
+import scala.slick.lifted.{BaseTypeMapper, Query}
+import com.keepit.eliza._
+import play.api.libs.functional.syntax._
+import play.api.libs.json.JsObject
+import com.keepit.common.mail.{GenericEmailAddress, EmailAddressHolder}
+import scala.slick.driver.BasicProfile
+import MessagingTypeMappers._
+import com.keepit.common.crypto.SimpleDESCrypt
+import com.keepit.social.{BasicNonUser, NonUserKinds, NonUserKind, BasicUserLikeEntity}
+
+sealed trait NonUserParticipant {
+  val identifier: String
+  val referenceId: Option[String]
+  val kind: NonUserKind
+
+  override def toString() = identifier.toString
+}
+object NonUserParticipant {
+  implicit val format = new Format[NonUserParticipant] {
+    // fields are shortened for overhead reasons
+    def reads(json: JsValue) = {
+      // k == "kind"
+      // i == "identifier"
+      // r == "referenceId"
+      ((json \ "k").asOpt[String], (json \ "i").asOpt[String]) match {
+        case (Some(NonUserKinds.email.name), Some(emailAddress)) =>
+          val addr = GenericEmailAddress(emailAddress)
+          val id = (json \ "r").asOpt[String].map(i => Id[EContact](i.toLong))
+          JsSuccess(NonUserEmailParticipant(addr, id))
+        case _ => JsError()
+      }
+    }
+    def writes(p: NonUserParticipant): JsValue = {
+      JsObject(Seq(Some("k" -> JsString(p.kind.name)), Some("i" -> JsString(p.identifier)), p.referenceId.map(r => "r" -> JsString(r))).flatten)
+    }
+  }
+
+  def toBasicNonUser(nonUser: NonUserParticipant) = {
+    // todo: Add other data, like econtact data
+    BasicNonUser(kind = nonUser.kind, id = nonUser.identifier, firstName = Some(nonUser.identifier), lastName = None)
+  }
+}
+
+case class NonUserEmailParticipant(address: EmailAddressHolder, econtactId: Option[Id[EContact]]) extends NonUserParticipant {
+  val identifier = address.address
+  val referenceId = econtactId.map(_.id.toString)
+  val kind = NonUserKinds.email
+}
+
+case class NonUserThread(
+  id: Option[Id[NonUserThread]] = None,
+  createdAt: DateTime = currentDateTime,
+  updatedAt: DateTime = currentDateTime,
+  participant: NonUserParticipant,
+  threadId: Id[MessageThread],
+  uriId: Option[Id[NormalizedURI]],
+  notifiedCount: Int,
+  lastNotifiedAt: Option[DateTime],
+  threadUpdatedAt: Option[DateTime],
+  muted: Boolean,
+  state: State[NonUserThread] = NonUserThreadStates.ACTIVE
+) extends Model[NonUserThread] {
+  def withId(id: Id[NonUserThread]): NonUserThread = this.copy(id = Some(id))
+  def withUpdateTime(updateTime: DateTime) = this.copy(updatedAt = updateTime)
+  def withState(state: State[NonUserThread]) = copy(state = state)
+}
+
+object NonUserThreadStates extends States[NonUserThread]
