+++ kifi-backend/modules/eliza/app/com/keepit/eliza/model/UriRenormalizationTrackingRepo.scala
@@ -1,4 +1,4 @@
-package com.keepit.eliza
+package com.keepit.eliza.model
 
 import com.keepit.common.db.{Model, Id}
 import com.keepit.common.db.slick.DBSession.{RWSession, RSession}
@@ -17,16 +17,16 @@ import org.joda.time.DateTime
 
 case class UriRenormalizationEvent(
     id: Option[Id[UriRenormalizationEvent]] = None,
-    createdAt: DateTime = currentDateTime, 
+    createdAt: DateTime = currentDateTime,
     updateAt: DateTime = currentDateTime,
     sequenceNumber: Long,
     numIdsChanged: Long,
     idsRetired: Seq[Id[NormalizedURI]]
-  ) 
+  )
   extends Model[UriRenormalizationEvent] {
 
   def withId(id: Id[UriRenormalizationEvent]): UriRenormalizationEvent = this.copy(id = Some(id))
-  def withUpdateTime(updateTime: DateTime) = this.copy(updateAt=updateTime) 
+  def withUpdateTime(updateTime: DateTime) = this.copy(updateAt=updateTime)
 }
 
 
@@ -43,8 +43,8 @@ trait UriRenormalizationTrackingRepo extends Repo[UriRenormalizationEvent] {
 }
 
 class UriRenormalizationTrackingRepoImpl @Inject() (
-    val clock: Clock, 
-    val db: DataBaseComponent 
+    val clock: Clock,
+    val db: DataBaseComponent
   ) extends DbRepo[UriRenormalizationEvent] with UriRenormalizationTrackingRepo {
